fix: removed some custom transition code since it was discouraged from Apple documentation
